refactoring #defines to typedefs

+++ src/Interface/gascoigne.h
@@ -48,7 +48,7 @@ typedef float MatrixEntryType;
 typedef double MatrixEntryType;
 #endif
 
-typedef size_t IndexType;                  // data type for global indices
+typedef unsigned long int IndexType;       // data type for global indices
 typedef unsigned short int ShortIndexType; // data type for local indices
 
 typedef nvector<double> GlobalParameterVector;
